@@ -31,51 +31,49 @@
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.15.5",
-    "@babel/plugin-transform-runtime": "^7.15.0",
-    "@babel/preset-env": "^7.15.6",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.15.0",
-    "@babel/runtime": "^7.15.4",
-    "@chiyu-git/eslint-config-react": "^1.0.8",
+    "@babel/core": "^7.18.6",
+    "@babel/plugin-transform-runtime": "^7.18.6",
+    "@babel/preset-env": "^7.18.6",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
+    "@babel/runtime": "^7.18.6",
+    "@chiyu-git/eslint-config-react": "^1.0.9",
     "@chiyu-git/stylelint-config-basic": "^1.0.3",
-    "@commitlint/cli": "^13.1.0",
-    "@commitlint/config-conventional": "^13.1.0",
+    "@commitlint/cli": "^17.0.3",
+    "@commitlint/config-conventional": "^17.0.3",
     "@types/animejs": "^3.1.4",
     "@types/chroma-js": "^2.1.3",
-    "@types/react": "^17.0.26",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.3.0",
-    "@typescript-eslint/eslint-plugin": "^4.30.0",
-    "@typescript-eslint/parser": "^4.30.0",
-    "babel-loader": "^8.2.2",
-    "conventional-changelog-cli": "^2.1.1",
-    "css-loader": "^6.3.0",
-    "eslint": "^7.32.0",
-    "fork-ts-checker-webpack-plugin": "^6.4.2",
-    "husky": "^3.1.0",
-    "less": "^4.1.1",
-    "less-loader": "^10.0.1",
-    "lint-staged": "^11.1.2",
-    "mini-css-extract-plugin": "^2.3.0",
-    "typescript": "^4.4.2",
-    "webpack": "^5.53.0",
-    "webpack-cli": "^4.8.0",
-    "webpack-dev-server": "^4.2.1",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.5",
+    "@types/react-router-dom": "^5.3.3",
+    "babel-loader": "^8.2.5",
+    "conventional-changelog-cli": "^2.2.2",
+    "css-loader": "^6.7.1",
+    "eslint": "^8.18.0",
+    "fork-ts-checker-webpack-plugin": "^7.2.11",
+    "husky": "3.1.0",
+    "less": "^4.1.3",
+    "less-loader": "^11.0.0",
+    "lint-staged": "^13.0.3",
+    "mini-css-extract-plugin": "^2.6.1",
+    "typescript": "^4.7.4",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.9.3",
     "webpack-merge": "^5.8.0"
   },
   "dependencies": {
     "animejs": "^3.2.1",
     "canvas2image": "^1.0.5",
     "chroma-js": "^2.4.2",
-    "core-js": "^3.18.1",
-    "echarts": "^5.2.1",
-    "echarts-for-react": "^3.0.1",
-    "html-webpack-plugin": "^5.3.2",
-    "html2canvas": "^1.3.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
-    "react-spring": "^9.3.2"
+    "core-js": "^3.23.3",
+    "echarts": "^5.3.3",
+    "echarts-for-react": "^3.0.2",
+    "html-webpack-plugin": "^5.5.0",
+    "html2canvas": "^1.4.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-spring": "^9.4.5"
   }
 }
