@@ -1,5 +1,7 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import Frame from './layouts/Frame';
 
-ReactDOM.render(<Frame />, document.querySelector('#root'));
+const container = document.querySelector('#root')
+const root = createRoot(container!); 
+root.render(<Frame  />);
