@@ -1,6 +1,6 @@
 // src/layouts/Frame.js
 import React from 'react';
-import { Switch, Route, BrowserRouter } from 'react-router-dom';
+import { Routes, Route, BrowserRouter } from 'react-router-dom';
 import Detail from '@src/components/Detail';
 import MemberInfo from '@src/components/MemberInfo/MemberInfo';
 import Slide from './Slide';
@@ -8,10 +8,9 @@ import Slide from './Slide';
 const Frame = () => (
     <BrowserRouter>
         <div>
-            <Switch>
-                <Route exact path = '/' component = {MemberInfo} />
-                <Route path = '/detail/:id' component = {Detail} />
-            </Switch>
+            <Routes>
+                <Route  path = '/' element = {<MemberInfo/>} />
+            </Routes>
             <Slide />
         </div>
     </BrowserRouter>
