@@ -5,7 +5,7 @@ import { Category } from '@common/root';
 export type MemberInfoListMap = {
     [Category.chara]: MemberInfoMap<Category.chara>;
     [Category.couple]: MemberInfoMap<Category.couple>;
-    [Category.seiyuu]: MemberInfoMap<Category.seiyuu>;
+    [Category.person]: MemberInfoMap<Category.person>;
 }
 
 export const initMemberInfoContext: MemberInfoListMap = {} as MemberInfoListMap;
