@@ -1,12 +1,12 @@
 import React, { FC, useContext } from 'react';
 import { Category } from '@common/root';
-import { SeiyuuRecordType } from '@common/record';
+import { PersonRecordType } from '@common/record';
 import { WeeklyContext } from '../weekly-context-manager';
 import IncrementRank from './IncrementRank';
 
 const TwitterFollowerWeekIncrementRank: FC<unknown> = () => {
-    const category = Category.seiyuu;
-    const recordType = SeiyuuRecordType.twitterFollower;
+    const category = Category.person;
+    const recordType = PersonRecordType.twitterFollower;
     const weeklyContext = useContext(WeeklyContext);
     const weeklyInfo = weeklyContext[category][recordType];
 
@@ -16,7 +16,7 @@ const TwitterFollowerWeekIncrementRank: FC<unknown> = () => {
             category = {category}
             recordType = {recordType}
             rankBarOptions = {{
-                title: 'pixiv-follower-周增榜',
+                title: 'twitter-follower-周增榜',
                 linearGradient: 'rgb(136,141,255)',
                 icon: `icon-${recordType.replace(/_/g, '-')}`,
             }}
