@@ -1,24 +1,24 @@
 import React, { createContext } from 'react';
 import { Category } from '@common/root';
-import { CharaRecordType, CoupleRecordType, SeiyuuRecordType } from '@common/record';
+import { CharaRecordType, CoupleRecordType, PersonRecordType } from '@common/record';
 import { RecordTypeWeeklyInfo } from '@common/weekly';
 
 type CharaRecordWeeklyInfo = Record<CharaRecordType, RecordTypeWeeklyInfo>
 
 type CoupleRecordWeeklyInfo = Record<CoupleRecordType, RecordTypeWeeklyInfo>
 
-type SeiyuuRecordWeeklyInfo = Record<SeiyuuRecordType, RecordTypeWeeklyInfo>;
+type PersonRecordWeeklyInfo = Record<PersonRecordType, RecordTypeWeeklyInfo>;
 
 export type WeeklyInfo = {
     [Category.chara]: CharaRecordWeeklyInfo;
     [Category.couple]: CoupleRecordWeeklyInfo;
-    [Category.seiyuu]: SeiyuuRecordWeeklyInfo;
+    [Category.person]: PersonRecordWeeklyInfo;
 };
 
 export const initWeeklyContext: WeeklyInfo = {
     [Category.chara]: {} as CharaRecordWeeklyInfo,
     [Category.couple]: {} as CoupleRecordWeeklyInfo,
-    [Category.seiyuu]: {} as SeiyuuRecordWeeklyInfo,
+    [Category.person]: {} as PersonRecordWeeklyInfo,
 };
 // createContext 参考
 // https://stackoverflow.com/questions/61333188/react-typescript-avoid-context-default-value
